% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHPMF-package.R
\docType{package}
\name{BHPMF-package}
\alias{BHPMF}
\alias{BHPMF-package}
\title{An R Package to fill missing values in a data matrix.}
\description{
An R package to filling gaps in a data matrix with hierarchical information (e.g. taxonomy) using Gibbs sampling (BHPMF).
}
\keyword{internal}
